requirements.txt
langchain-openai>=0.1.0
langgraph>=0.2.0
langchain-core>=0.2.0
chromadb>=0.5.0
sentence-transformers>=2.7.0
openai>=1.30.0
pyyaml>=6.0
jinja2>=3.1.4
requests>=2.31.0
python-dotenv>=1.0.0
rich>=13.0.0
playwright>=1.44.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
